Improve interface

+++ src/main/java/coffeeshop/controller/AdminProductController.java
@@ -41,6 +41,7 @@ import coffeeshop.resource.product.ProductDetailResource;
 import coffeeshop.resource.product.ProductRegistResource;
 import coffeeshop.resource.product.ProductResource;
 import coffeeshop.resource.product.ProductUpdateResource;
+import coffeeshop.service.OrderService;
 import coffeeshop.service.product.ProductService;
 
 @Controller
@@ -54,7 +55,8 @@ public class AdminProductController extends BaseController {
 	private MessageSource messageSource;
 	@Autowired
 	private ProductService productService;
-
+	@Autowired
+	private OrderService orderService;
 	@Autowired
 	private ProductHelper productHelper;
 
@@ -138,7 +140,6 @@ public class AdminProductController extends BaseController {
 
 		// Validate
 		if (result.hasErrors()) {
-			redirectAttributes.addFlashAttribute("error", "Error ...");
 			return viewPrefix + "create_product";
 		}
 		String imageUrl;
@@ -146,7 +147,6 @@ public class AdminProductController extends BaseController {
 		try {
 			imageUrl = this.doUploadFile(resource.getImage());
 		} catch (IOException e) {
-			redirectAttributes.addFlashAttribute("error", "Error ...");
 			return viewPrefix + "create_product";
 		}
 
@@ -193,7 +193,7 @@ public class AdminProductController extends BaseController {
 	@PatchMapping("/{productId}")
 	public String updateProduct(@PathVariable("productId") Integer productId,
 			@Valid @ModelAttribute("product") ProductUpdateResource resource, BindingResult result, Model model,
-			RedirectAttributes redirectAttributes) {
+			RedirectAttributes redirectAttributes, Locale locale) {
 
 		// Validate
 		if (result.hasErrors()) {
@@ -215,32 +215,39 @@ public class AdminProductController extends BaseController {
 
 		// return view
 		redirectAttributes.addAttribute("productId", productId);
-		redirectAttributes.addFlashAttribute("info", "Product updated successfully");
+		redirectAttributes.addFlashAttribute("info", messageSource.getMessage("info.product.updated", null, locale));
 		return "redirect:/" + viewPrefix + "{productId}";
 	}
 
 	@DeleteMapping("/{productId}")
 	public String deleteProduct(@PathVariable("productId") Integer productId, Model model,
-			RedirectAttributes redirectAttributes) {
+			RedirectAttributes redirectAttributes, Locale locale) {
 
 		// 商品のアクセス権限をチェック
 		if (!productService.existProduct(productId)) {
 			// return forbidden view
 			return "403";
 		}
-
+		Product product = productService.getProductDetail(productId);
+		if(orderService.checkIfProductIsInActiveOrder(product)){
+			ProductDetailResource resource = productHelper.createProductDetailResource(product);
+			model.addAttribute("info", messageSource.getMessage("error.delete.product", null, locale));
+			model.addAttribute("product", resource);
+			return viewPrefix + "product";
+		}
+		
 		// 商品削除
 		productService.deleteProduct(productId);
 
 		// return view
-		redirectAttributes.addFlashAttribute("info", "Product deleted successfully");
+		redirectAttributes.addFlashAttribute("info", messageSource.getMessage("info.product.deleted", null, locale));
 		return "redirect:/" + viewPrefix;
 	}
 
 	@PatchMapping("/{productId}/updateImage")
 	@ResponseBody
 	public ResponseEntity<?> uploadFile(@PathVariable("productId") Integer productId,
-			@RequestParam("image") MultipartFile image) {
+			@RequestParam("image") MultipartFile image, Locale locale) {
 
 		if (image.isEmpty()) {
 			return new ResponseEntity<>("Please select a image!", HttpStatus.BAD_REQUEST);
@@ -266,7 +273,7 @@ public class AdminProductController extends BaseController {
 		}
 		HashMap<String, Object> resultMap = new HashMap<String, Object>();
 		resultMap.put("imageUrl", image.getOriginalFilename());
-		resultMap.put("message", "Product image updated successfully");
+		resultMap.put("message", messageSource.getMessage("info.product.image.updated", null, locale));
 		return new ResponseEntity<HashMap<String, Object>>(resultMap, new HttpHeaders(), HttpStatus.OK);
 	}
